server/requirements.txt
notebook==6.4.12
numpy==1.23.1
pandas==1.4.3
psycopg2-binary==2.9.3
pyserum==0.5.0a0
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
solana==0.25.0
SQLAlchemy==1.4.39
fastapi==0.79.0
uvicorn==0.18.2
fastapi-utils==0.2.1